@@ -1,30 +1,31 @@
 // Code generated by Wire. DO NOT EDIT.
 
 //go:generate go run github.com/google/wire/cmd/wire
-//+build !wireinject
+//go:build !wireinject
+// +build !wireinject
 
 package di
 
 import (
-	"will/app/controller"
-	"will/app/daemon_jobs/cron"
-	"will/app/modules/mysql/dao"
-	"will/app/router"
-	"will/app/service"
-	"will/utils/validator"
+	"willshark/app/controller"
+	"willshark/app/daemon_jobs/cron"
+	"willshark/app/modules/mysql/dao"
+	"willshark/app/router"
+	"willshark/app/service"
+	"willshark/utils/validator"
 )
 
 // Injectors from wire.go:
 
 func BuildInjector() (*App, func(), error) {
 	validatorX := validator.NewValidator()
-	redisPool, cleanup := InitRedis()
+	redis, cleanup := InitRedis()
 	mysqlPool, cleanup2 := InitMysql()
 	user := dao.User{
 		MysqlPool: mysqlPool,
 	}
 	apps := service.Apps{
-		Rds:  redisPool,
+		Rds:  redis,
 		User: user,
 	}
 	controllerApps := &controller.Apps{
@@ -39,7 +40,7 @@ func BuildInjector() (*App, func(), error) {
 	}
 	engine, cleanup3 := InitGinEngine(routers)
 	jobs := &cron.Jobs{
-		Rds:  redisPool,
+		Rds:  redis,
 		User: user,
 	}
 	app := &App{
