@@ -3,16 +3,18 @@ package middlewares
 import (
 	"context"
 	"github.com/gin-gonic/gin"
-	"will/consts"
-	"will/core"
+	"willshark/consts"
+	"willshark/core"
 )
 
 func StartTrace() gin.HandlerFunc {
 	return func(c *gin.Context) {
-		tr := core.Trace.Tracer(consts.JaegerRouter)
-		ctxRoot, span := tr.Start(context.Background(), c.Request.URL.Path)
-		c.Set(consts.JaegerContext, ctxRoot)
-		defer span.End()
-		c.Next()
+		if core.Trace != nil {
+			tr := core.Trace.Tracer(consts.JaegerRouter)
+			ctxRoot, span := tr.Start(context.Background(), c.Request.URL.Path)
+			c.Set(consts.JaegerContext, ctxRoot)
+			defer span.End()
+			c.Next()
+		}
 	}
 }
