@@ -7,10 +7,9 @@ import (
 	"strconv"
 	"sync/atomic"
 	"time"
-	"will/app/modules/redis"
-	"will/core"
-	"will/will_tools/logs"
-	"will/will_tools/strings"
+	"willshark/app/modules/redis"
+	"willshark/utils/logs/logger"
+	"willshark/utils/strings"
 )
 
 const (
@@ -20,7 +19,7 @@ const (
 )
 
 type RedisLock struct {
-	store   *redis.RedisPool
+	store   *redis.Redis
 	seconds uint32
 	key     string
 	id      string
@@ -30,7 +29,7 @@ func init() {
 	rand.Seed(time.Now().UnixNano())
 }
 
-func NewRedisLock(store *redis.RedisPool, key string) *RedisLock {
+func NewRedisLock(store *redis.Redis, key string) *RedisLock {
 	return &RedisLock{
 		store: store,
 		key:   key,
@@ -48,15 +47,12 @@ func (rl *RedisLock) Acquire() (bool, error) {
 
 func (rl *RedisLock) AcquireCtx(ctx context.Context) (bool, error) {
 	seconds := atomic.LoadUint32(&rl.seconds)
-	resp, err := rl.store.EvalCtx(ctx, lockCommand,
+	resp, err := rl.store.Eval(lockCommand,
 		[]string{rl.key}, rl.id, strconv.Itoa(int(seconds)*millisPerSecond+tolerance))
 
 	if err != nil {
 		// Todo: when do lock_test annotation the log plugin
-		AcquireCtxLog := logs.StringFormatter{
-			Msg: fmt.Sprintf("Error on acquiring lock for %s, %s", rl.key, err.Error()),
-		}
-		_ = core.Log.Error(AcquireCtxLog)
+		logger.Error(fmt.Sprintf("Error on acquiring lock for %s, %s", rl.key, err.Error()))
 		return false, err
 	} else if resp == nil {
 		return false, nil
@@ -68,10 +64,7 @@ func (rl *RedisLock) AcquireCtx(ctx context.Context) (bool, error) {
 	}
 
 	// Todo: when do lock_test annotation the log plugin
-	AcquireCtxLog := logs.StringFormatter{
-		Msg: fmt.Sprintf("Unknown reply when acquiring lock for %s: %v", rl.key, resp),
-	}
-	_ = core.Log.Error(AcquireCtxLog)
+	logger.Error(fmt.Sprintf("Unknown reply when acquiring lock for %s: %v", rl.key, resp))
 
 	return false, nil
 }
@@ -81,12 +74,9 @@ func (rl *RedisLock) Release() (bool, error) {
 }
 
 func (rl *RedisLock) ReleaseCtx(ctx context.Context) (bool, error) {
-	resp, err := rl.store.EvalCtx(ctx, delCommand, []string{rl.key}, rl.id)
+	resp, err := rl.store.Eval(delCommand, []string{rl.key}, rl.id)
 	if err != nil {
-		ReleaseCtxLog := logs.StringFormatter{
-			Msg: redis.ReleaseCtxErr.Error() + err.Error(),
-		}
-		_ = core.Log.Error(ReleaseCtxLog)
+		logger.Error(redis.ReleaseCtxErr.Error() + err.Error())
 
 		return false, err
 	}
