@@ -8,19 +8,36 @@ import (
 	"go.opentelemetry.io/otel/sdk/resource"
 	sdktrace "go.opentelemetry.io/otel/sdk/trace"
 	semconv "go.opentelemetry.io/otel/semconv/v1.10.0"
-	"will/consts"
+	"willshark/consts"
+	"willshark/utils/logs/logger"
 )
 
 var Trace *sdktrace.TracerProvider
 
+type jeagerConfig struct {
+	Host    string `yaml:"host,omitempty"`
+	Enables bool   `yaml:"enable,omitempty"`
+}
+
 func initJaeger() func() {
-	Trace = initConn(CoreConfig["jaegerhost"].(string)+consts.JaegerURL, consts.APP_NAME+"_"+consts.APP_ID)
-	return func() {
-		ctx, cancelFunc := context.WithCancel(context.Background())
-		Trace.Shutdown(ctx)
-		cancelFunc()
-		Log.SuccessDefault("Close Jaeger Sunccess")
+	cfg, err := GetSingleConfig(CoreConfig, "jaeger", jeagerConfig{})
+	if err != nil {
+		panic("init jeager failed, jeager config not found")
 	}
+	jConfig, ok := cfg.(*jeagerConfig)
+	if !ok {
+		panic("init jeager failed, jeager config incorrect")
+	}
+	if jConfig.Enables {
+		Trace = initConn(jConfig.Host+consts.JaegerURL, consts.APP_NAME+"_"+consts.APP_ID)
+		return func() {
+			ctx, cancelFunc := context.WithCancel(context.Background())
+			Trace.Shutdown(ctx)
+			cancelFunc()
+			logger.Info("Close Jaeger Sunccess")
+		}
+	}
+	return func() {}
 }
 
 func initConn(host, serviceName string) *sdktrace.TracerProvider {
