@@ -1,14 +1,17 @@
 package core
 
 import (
+	"context"
 	"errors"
-	"github.com/gomodule/redigo/redis"
 	"time"
+	"willshark/utils/logs/logger"
+
+	"github.com/redis/go-redis/v9"
 )
 
 var (
 	RedisConfig map[string]interface{}
-	RedisPool   map[string]*redis.Pool
+	RedisPool   map[string]*redis.Client
 )
 
 type RedisConn struct {
@@ -25,65 +28,79 @@ type RedisConn struct {
 func initRedis() func() {
 	RedisConfig, err := GetMapConfig(CoreConfig, "redis", RedisConn{})
 	if err != nil {
-		return func() {
-		}
+		return func() {}
 	}
 	if len(RedisConfig) < 1 {
 		panic("init redis pool config failed, redis config not found")
 	}
-	RedisPool = make(map[string]*redis.Pool)
+	RedisPool = make(map[string]*redis.Client)
 	for name, val := range RedisConfig {
-		if rdb, _ := initRedisPool(val.(*RedisConn)); rdb != nil {
+		rdb, err := initRedisClient(val.(*RedisConn))
+		if err != nil {
+			panic("initRedisClient err: " + err.Error())
+		}
+		if rdb != nil {
 			RedisPool[name] = rdb
 		}
 	}
 
 	return func() {
-		for key, value := range RedisPool {
-			if err := value.Close(); err != nil {
-				_ = Log.PanicDefault("Redis[ " + key + " ]Close Err:" + err.Error())
+		for key, client := range RedisPool {
+			if err := client.Close(); err != nil {
+				logger.Error("Redis[ " + key + " ]Close Err:" + err.Error())
 				continue
 			}
-			_ = Log.SuccessDefault("Redis[ " + key + " ]Close Success!")
+			logger.Info("Redis[ " + key + " ]Close Success!")
 		}
-
 	}
 }
 
-func initRedisPool(conne *RedisConn) (*redis.Pool, error) {
-	redisPool := &redis.Pool{
-		MaxIdle:     conne.MaxIdleNum,
-		MaxActive:   conne.MaxActive,
-		IdleTimeout: time.Duration(conne.MaxIdleTimeout) * time.Second,
-		Wait:        true,
-		Dial: func() (redis.Conn, error) {
-			con, err := redis.Dial("tcp", conne.Host,
-				redis.DialPassword(conne.PassWord), redis.DialDatabase(conne.DataBase),
-				redis.DialConnectTimeout(time.Duration(conne.ConnectTimeout)*time.Second),
-				redis.DialReadTimeout(time.Duration(conne.ReadTimeout)*time.Second),
-				redis.DialWriteTimeout(time.Duration(conne.ReadTimeout)*time.Second),
-			)
-			if err != nil {
-				return nil, err
-			}
-			return con, nil
-		},
+func initRedisClient(conf *RedisConn) (*redis.Client, error) {
+	options := &redis.Options{
+		Addr:            conf.Host,
+		Password:        conf.PassWord,
+		DB:              conf.DataBase,
+		MaxIdleConns:    conf.MaxIdleNum,
+		PoolSize:        conf.MaxActive,
+		ConnMaxIdleTime: time.Duration(conf.MaxIdleTimeout) * time.Second,
+		DialTimeout:     time.Duration(conf.ConnectTimeout) * time.Second,
+		ReadTimeout:     time.Duration(conf.ReadTimeout) * time.Second,
+		WriteTimeout:    time.Duration(conf.ReadTimeout) * time.Second,
+
+		// 自定义配置
+		MinIdleConns:    conf.MaxIdleNum / 4,
+		PoolTimeout:     5 * time.Second,
+		ConnMaxLifetime: time.Hour,
+		MaxRetries:      3,
+		MinRetryBackoff: 8 * time.Millisecond,
+		MaxRetryBackoff: 512 * time.Millisecond,
 	}
-	return redisPool, nil
+
+	client := redis.NewClient(options)
+
+	// 测试连接
+	ctx := context.Background()
+	if err := client.Ping(ctx).Err(); err != nil {
+		return nil, err
+	}
+
+	return client, nil
 }
 
-func GetRedisDB(key string) (rdb *redis.Pool, err error) {
-	rdbPool, ok := RedisPool[key]
+func GetRedisDB(key string) (*redis.Client, error) {
+	client, ok := RedisPool[key]
 	if !ok {
-		if config, ok := RedisConfig[key]; !ok {
+		config, exists := RedisConfig[key]
+		if !exists {
 			return nil, errors.New(key + " redis dbConfig doesn't exist")
-		} else {
-			if rdbPool, err := initRedisPool(config.(*RedisConn)); err != nil {
-				return nil, errors.New(key + " redis dbConfig Initialization failure")
-			} else {
-				RedisPool[key] = rdbPool
-			}
 		}
+
+		newClient, err := initRedisClient(config.(*RedisConn))
+		if err != nil {
+			return nil, errors.New(key + " redis dbConfig Initialization failure")
+		}
+		RedisPool[key] = newClient
+		return newClient, nil
 	}
-	return rdbPool, nil
+	return client, nil
 }
